[FE][S13P21D108-127] fix: 환전 예약 로직 오류 수정

+++ frontend/app/src/main/java/com/d108/moyeo/data/remote/dto/exchange/CreateReservationRequestDto.kt
@@ -12,7 +12,7 @@ data class CreateReservationRequestDto(
     @SerializedName("amount")
     val amount: Long,
     @SerializedName("targetRate")
-    val targetRate: Double,
+    val targetRate: Long,
     @SerializedName("expiresAt")
     val expiresAt: String,
 ){
